@@ -41,94 +42,92 @@ sphinx-autoapi = "^3.0.0"
 sphinx-rtd-theme = "^2.0.0"
 
 [tool.coverage.run]
-source = ['chaturbate_poller']
+source = ['chaturbate_event_listener']
 relative_files = true
 
 [tool.semantic_release]
 version_variable = "pyproject.toml:version"
 version_toml = [
-    "pyproject.toml:tool.poetry.version",
-]                                                    # version location
-branch = "main"                                      # branch to make releases of
-changelog_file = "CHANGELOG.md"                      # changelog file
-build_command = "pip install poetry && poetry build" # build dists
+  "pyproject.toml:tool.poetry.version",
+]
+branch = "main"
+changelog_file = "CHANGELOG.md"
+build_command = "pip install poetry && poetry build"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
+[tool.pytest.ini_options]
+addopts = "--strict-markers"
+generate_report_on_test = "True"
+asyncio_mode = "strict"
+
 [tool.ruff]
-# Exclude a variety of commonly ignored directories.
 exclude = [
-    ".bzr",
-    ".direnv",
-    ".eggs",
-    ".git",
-    ".git-rewrite",
-    ".hg",
-    ".ipynb_checkpoints",
-    ".mypy_cache",
-    ".nox",
-    ".pants.d",
-    ".pyenv",
-    ".pytest_cache",
-    ".pytype",
-    ".ruff_cache",
-    ".svn",
-    ".tox",
-    ".venv",
-    ".vscode",
-    "__pypackages__",
-    "_build",
-    "buck-out",
-    "build",
-    "dist",
-    "node_modules",
-    "site-packages",
-    "venv",
+  ".bzr",
+  ".direnv",
+  ".eggs",
+  ".git",
+  ".git-rewrite",
+  ".hg",
+  ".ipynb_checkpoints",
+  ".mypy_cache",
+  ".nox",
+  ".pants.d",
+  ".pyenv",
+  ".pytest_cache",
+  ".pytype",
+  ".ruff_cache",
+  ".svn",
+  ".tox",
+  ".venv",
+  ".vscode",
+  "__pypackages__",
+  "_build",
+  "buck-out",
+  "build",
+  "dist",
+  "node_modules",
+  "site-packages",
+  "venv",
+  "out",
+  "docs"
 ]
-
-# Same as Black.
-line-length = 88
+src = ["src", "tests", "docs"]
+line-length = 100
 indent-width = 4
-
-# Assume Python 3.10
-target-version = "py310"
+target-version = "py311"
+extend-include = ["*.ipynb"]
 
 [tool.ruff.lint]
 select = ["ALL"]
 ignore = [
-    "COM812", # missing-trailing-comma
-    "ISC001", # single-line-implicit-string-concatenation
-    "ANN101", "ANN102", # missing-type-self (redundant)
-    "D203", # one-blank-line-before-class (incompatable)
-    "D213", # multi-line-summary-second-line (incompatable)
-    "S101" # disable assert warning for tests
+  "COM812",  # conflicts with Ruff formatter
+  "ISC001",  # conflicts with Ruff formatter
+  "ANN101",  # missing-type-self (depricated)
+  "ANN102",  # missing-type-cls (depricated)
 ]
-# Allow fix for all enabled rules (when `--fix`) is provided.
 fixable = ["ALL"]
 unfixable = []
-
-# Allow unused variables when underscore-prefixed.
-dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
 
 [tool.ruff.lint.pydocstyle]
 convention = "google"
 
 [tool.ruff.format]
-# Like Black, use double quotes for strings.
 quote-style = "double"
-
-# Like Black, indent with spaces, rather than tabs.
 indent-style = "space"
-
-# Like Black, respect magic trailing commas.
 skip-magic-trailing-comma = false
-
-# Like Black, automatically detect the appropriate line ending.
 line-ending = "auto"
-
-# Enable auto-formatting of code examples in docstrings. Markdown,
-# reStructuredText code/literal blocks and doctests are all supported.
 docstring-code-format = true
 docstring-code-line-length = "dynamic"
+
+[tool.mypy]
+exclude = ["out"]
+disable_error_code = "import-untyped"
+
+[tool.bandit]
+exclude_dirs = ["tests",".venv"]
+
+[tool.poetry.scripts]
+chaturbate_event_listener = "chaturbate_event_listener.__main__:start"
