@@ -1,9 +1,12 @@
-import asyncio  # noqa: INP001, D100
+"""Example of using the Chaturbate Poller to fetch events."""
+
+import asyncio
 import logging
 import os
 
-from chaturbate_poller import ChaturbateClient
 from dotenv import load_dotenv
+
+from chaturbate_poller import ChaturbateClient
 
 load_dotenv()
 
@@ -11,7 +14,8 @@ username = os.getenv("CB_USERNAME", "")
 token = os.getenv("CB_TOKEN", "")
 
 
-async def main() -> None:  # noqa: D103
+async def main() -> None:
+    """Run the main function."""
     async with ChaturbateClient(
         username, token, 20, "https://events.testbed.cb.dev/events/{username}/{token}/"
     ) as client:
